[#44]feat: 함수분리해서 폴더구조 정리
