+++ src/component/RightSection/ForBox.tsx
