+++ src/component/RightSection/VariableBox.module.css
