+++ src/component/RightSection/VariableBox.tsx
