+++ src/component/RightSection/utils/turnLight.ts
@@ -1,0 +1,28 @@
+import { ActivateItem } from "@/types/activateItem";
+import { AllObjectItem } from "@/types/allObjectItem";
+
+// 하이라이트 효과를 줘야하는 부분을 표시해주는 함수
+export const turnLight = (
+  codeFlows: AllObjectItem[], //비주얼 스택
+  Activate: ActivateItem[] //활성화 스택
+): AllObjectItem[] => {
+  return codeFlows.map((codeFlow) => {
+    if (Activate.some((data) => data.id === codeFlow.id)) {
+      // 수정해야하는 위치일때
+      // isLight을 true로 바꾼다
+      return {
+        ...codeFlow,
+        isLight: true,
+        child: turnLight(codeFlow.child, Activate),
+      };
+    } else if (codeFlow.child && codeFlow.child.length > 0) {
+      return {
+        ...codeFlow,
+        isLight: false,
+        child: turnLight(codeFlow.child, Activate),
+      };
+    } else {
+      return { ...codeFlow, isLight: false };
+    }
+  });
+};
