--- src/component/Visualizer.module.css
@@ -1,15 +1,0 @@
-.container {
-  overflow: scroll;
-  width: 100%;
-  height: 80vh;
-  border: 1px solid #ccc;
-}
-
-.box {
-  width: "50px";
-  height: "50px";
-  background-color: #4caf50;
-  display: "flex";
-  justify-content: "center";
-  align-items: "center";
-}
